test_mqtt_automated/requirements.txt
pymongo
paho-mqtt
certifi
dnspython